Init with updated .gitignore
