+++ Api/ProjectAutism/Credential.cs
@@ -1,0 +1,12 @@
+﻿namespace ProjectAutism;
+
+public class Credential
+{
+    public string Email {
+        get;
+        set;
+    }
+
+    public string Name { get; set; }
+    public string Surname { get; set; }
+}
