+++ Api/ProjectAutism/Repos/IGatheringRepository.cs
@@ -1,10 +1,11 @@
 ﻿using ProjectAutism.Data.Models;
 
-namespace ProjectAutism.Data.Repos;
+namespace ProjectAutism.Repos;
 
 public interface IGatheringRepository
 {
     IEnumerable<Gathering> GetGatherings();
     Gathering CreateGathering(Gathering gathering);
     bool DeleteGathering(int id);
+    bool SubscribeToGathering(Credential credential, int gatheringId);
 }
