+++ Api/ProjectAutism/appsettings.Development.json
@@ -4,5 +4,12 @@
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning"
     }
+  },
+  "EmailSettings": {
+    "SmtpServer": "smtp.gmail.com",
+    "SmtpPort": 587,
+    "Username": "s93126816@gmail.com",
+    "Password": "TestCSharp2"
   }
 }
+
